removed personal projects
