+++ keyboards/fastbord/config.h
@@ -25,12 +25,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define PRODUCT_ID      0x6060
 #define DEVICE_VER      0x0001
 #define MANUFACTURER    You
-#define PRODUCT         MAX96
+#define PRODUCT         fastbord
 #define DESCRIPTION     A custom keyboard
 
 /* key matrix size */
-#define MATRIX_ROWS 5
-#define MATRIX_COLS 18
+#define MATRIX_ROWS 4
+#define MATRIX_COLS 12
 
 /*
  * Keyboard Matrix Assignments
@@ -42,8 +42,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *                  ROW2COL = ROW = Anode (+), COL = Cathode (-, marked on diode)
  *
 */
-#define MATRIX_ROW_PINS { B3, B2, B1, B0, D3 }
-#define MATRIX_COL_PINS { F7, F6, F5, F4, F1, F0, E6, E2, D7, D6, D5, D4, C7, C6, B7, B6, B5, B4  }
+#define MATRIX_ROW_PINS { B1, B3, B2, B6 }
+#define MATRIX_COL_PINS { B5, B4, E6, D7, C6, D4, D0, D1, F4, F5, F6, F7 }
 #define UNUSED_PINS
 
 /* COL2ROW or ROW2COL */
@@ -52,12 +52,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // #define BACKLIGHT_PIN B7
 // #define BACKLIGHT_BREATHING
 // #define BACKLIGHT_LEVELS 3
-#define RGB_DI_PIN D2
-#define RGBLIGHT_ANIMATIONS
-#define RGBLED_NUM 16    // Number of LEDs
-#define RGBLIGHT_HUE_STEP 10
-#define RGBLIGHT_SAT_STEP 17
-#define RGBLIGHT_VAL_STEP 17
+
 
 /* Debounce reduces chatter (unintended double-presses) - set 0 if debouncing is not needed */
 #define DEBOUNCING_DELAY 5
