+++ keyboards/fastbord/keymaps/default/readme.md
@@ -1,0 +1,1 @@
+# The default keymap for fastbord
